pyproject.toml
[tool.poetry]
name = "pynace"
version = "0.1.0"
description = ""
authors = ["Mohammed Salman <mohammed@holvi.com>"]

[tool.poetry.dependencies]
python = "^2.7 || ^3.6 ||  ^3.7"
requests = "^2.26.0"
attrs = "^21.2.0"
six = "^1.16.0"
csv342 = "^1.0.1"
typing = {version = "^3.10.0", python = "2.7"}

[tool.poetry.dev-dependencies]
pytest = [
    { version = "^5.2", python = "^3.4" },
    { version = "~4.6", python = "~2.7" }
]
ipdb = "^0.13.9"
pylama = "^7.7.1"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
